pyproject.toml
[project]
name = "koyeb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "koyeb-sdk>=1.2.0",
    "langchain-anthropic>=1.1.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
